production_requirements.txt
python-dotenv
numpy
rdkit
plotly
streamlit
pyvis
arango
python-arango
langchain
langchain-core
langchain-community
langchain-openai
openai 
